ContainerAdapterFactory -> ActivationContext changes for ContainerTests
