+++ Rebus.CastleWindsor.Tests/CastleWindsorRealContainerTests.cs
@@ -4,7 +4,7 @@ using Rebus.Tests.Contracts.Activation;
 namespace Rebus.CastleWindsor.Tests
 {
     [TestFixture]
-    public class CastleWindsorRealContainerTests : RealContainerTests<CastleWindsorContainerAdapterFactory>
+    public class CastleWindsorRealContainerTests : RealContainerTests<CastleWindsorActivationContext>
     {
     }
 }
